style: better style
