+++ unity-renderer/Assets/DCLPlugins/ECS7/ECSComponents/AvatarModifierArea/Tests/DCL.ECSComponents.AvatarModifierArea.Tests.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "DCL.ECSComponents.AvatarModifierAreaShould",
+    "name": "DCL.ECSComponents.AvatarModifierArea.Tests",
     "rootNamespace": "",
     "references": [
         "GUID:b1087c5731ff68448a0a9c625bb7e52d",
